+++ commands/__init__.py
@@ -1,0 +1,5 @@
+import os
+__all__ = [os.path.splitext(x)[0] for x in os.listdir(__name__)
+           if os.path.splitext(x)[1] == ".py"
+           and not x.startswith("__")]
+from . import *
