+++ commands/ping.py
@@ -1,0 +1,5 @@
+NAME = "ping"
+DESCRIPTION = "Prints pong"
+
+async def run(ctx):
+    await ctx.respond("pong")
