+++ constants.py
@@ -1,0 +1,16 @@
+OUR_TEAM = 293
+IMAGE_DIR = "images"
+EVENT_KEYS = ["njski", "njrob"]
+OWNER_IDS = [425618103062364160, 708440911591243826]
+
+TEAM_COLORS = [
+    "#000000",
+    "#ff0000",
+    "#0000ff",
+    "#00ffff",
+    "#ff00ff",
+    "#cc00ff",
+    "#ff8800",
+    "#888888",
+    "#44cc44"
+]
