+++ utils.py
@@ -4,6 +4,8 @@ import hikari
 import os
 import random
 import string
+import statbotics
+from subsystems.Cache import Cache
 
 
 IS_REPLIT = os.getenv("IS-REPLIT")
@@ -27,3 +29,19 @@ async def send_plot(ctx: lightbulb.Context, title: str) -> None:
     embed.set_image(hikari.File(path))
     await ctx.respond(embed)
     os.remove(path)
+
+cache = Cache()
+
+def get_team_year(team_number, year):
+    key = f"teamyear/{team_number},{year}"
+    if cache.is_outdated(key):
+        try: cache[key] = sb.get_team_year(team_number, year)
+        except UserWarning: cache[key] = None
+    return cache[key]
+
+
+async def error(ctx, text):
+    await ctx.respond(hikari.Embed(title = "Error", description = text, color = "#ff0000"))
+
+
+sb = statbotics.Statbotics()
